Create bin_except.py
project-4a
Modify the binary search function from the exploration so that, instead of returning -1 when the target value is not in the list, raises a TargetNotFound exception (you'll need to define this exception class). Otherwise it should function normally. Name this function bin_except.

The file must be named: bin_except.py